test: Fix T350-crypto.sh on Emacs snapshots
The "reply to encrypted message" subtest of T350-crypto.sh was failing
with Emacs snapshots because it assumed the "In-Reply-To" and
"References" headers it attempts to filter out would never be
line-wrapped. This apparently changed in the latest Emacs versions,
which causes the test to fail when ran against them.

* T350-crypto.sh: Unwrap headers before filtering them.